Fix warnings on Visual Studio (#1428)
* Fix warning C4244 in src/layer/normalize.cpp

C4244: '=': conversion from 'double' to 'float', possible loss of data

* Fix warning C4244 in src/layer/requantize.cpp

C4244: 'initializing': conversion from 'double' to 'int', possible loss of data

* Fix warning C4244 in src/mat_pixel_resize.cpp

C4244: '=': conversion from 'double' to 'int', possible loss of data

* Fix warning C4244 in src/mat_pixel.cpp

C4244: '=': conversion from 'int' to 'float', possible loss of data

* Fix warning C4267 in src/modelbin.cpp

C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

* Fix warning C4244 in src/layer/batchnorm.cpp

C4244: 'initializing': conversion from 'double' to 'float',
    possible loss of data

* Fix warning C4244 in src/layer/padding.cpp

C4244: 'argument': conversion from 'const float' to 'T', possible loss of data

* Fix warning C4244 in src/layer/priorbox.cpp

C4244: '=': conversion from 'double' to 'float',
    possible loss of data
C4244: 'initializing': conversion from 'double' to 'float',
    possible loss of data

+++ src/layer/batchnorm.cpp
@@ -60,7 +60,7 @@ int BatchNorm::load_model(const ModelBin& mb)
 
     for (int i=0; i<channels; i++)
     {
-        float sqrt_var = sqrt(var_data[i] + eps);
+        float sqrt_var = static_cast<float>(sqrt(var_data[i] + eps));
         a_data[i] = bias_data[i] - slope_data[i] * mean_data[i] / sqrt_var;
         b_data[i] = slope_data[i] / sqrt_var;
     }
